Merge from master

+++ src/Illuminate/Database/Console/Migrations/RefreshCommand.php
@@ -1,9 +1,12 @@
 <?php namespace Illuminate\Database\Console\Migrations;
 
 use Illuminate\Console\Command;
+use Illuminate\Console\ConfirmableTrait;
 use Symfony\Component\Console\Input\InputOption;
 
 class RefreshCommand extends Command {
+
+	use ConfirmableTrait;
 
 	/**
 	 * The console command name.
@@ -26,6 +29,8 @@ class RefreshCommand extends Command {
 	 */
 	public function fire()
 	{
+		if ( ! $this->confirmToProceed()) return;
+
 		$database = $this->input->getOption('database');
 
 		$this->call('migrate:reset', array('--database' => $database));
@@ -73,6 +78,8 @@ class RefreshCommand extends Command {
 	{
 		return array(
 			array('database', null, InputOption::VALUE_OPTIONAL, 'The database connection to use.'),
+
+			array('force', null, InputOption::VALUE_NONE, 'Force the operation to run when in production.'),
 
 			array('seed', null, InputOption::VALUE_NONE, 'Indicates if the seed task should be re-run.'),
